Fixed some operations in complex.gi so they use the AbelianCat interface.
Many operations in complex.gi (as it was written in QPA) handled
objects in a category by calling other functions than those included
in the Cat interface.

I have removed references to Source, Range, Kernel, Image and the
division operator (/).

HomologyOfComplex, which did not work in the earlier implementation,
should work now.