Made InfListString more flexible.
Allow more customizing of how to make string representations of
infinite lists.  In particular, the "opt" argument (a record)
specifies how to format the string.  The intention is that the
InfListString operations (which are implemented for all the different
types of InfList implementations) can be used for printing complexes,
just by making an "opt" record which specifies the format for printing
complexes, and setting the other arguments to appropriate values.