Started implementing \= for inflists.
Tried implementing \= for inflists by comparing the implementation
objects.  This turns out to not work well, since the implementation
may change over time, and thus equal objects may become different, or
different objects may become equal.

This probably means that I have to do some radical changes to the
inflist structures.  The code in this commit is not useful now, but I
commit it anyway in case some of it will turn out to be useful after I
have made the necessary changes to the data structures.