Restructured inflists to make equality comparisons possible.
Inflists now remember (to a certain extent) how they were created.
For example, an inductive list will always stay inductive, even if it
finds out that it is equivalent to a repeating list.  When it
discovers this, it creates a repeating list which it uses for lookup,
but it continues to use its own information for equality comparisons.

The former distinction between "inflists" and "inflist
implementations" is therefore gone.  The inflist objects themselves
contain all the information that is needed to describe the inflist,
but they may have a separate implementation object (which is also an
inflist) to use for lookup.

I have tried to implement equality functions.  I'm not sure if they
work correctly in all cases yet (probably not), but I think it should
be possible to make them work in a useful and consistent manner.